FFM-7275 - iOS SDK - SSE event stream doesn't time out when no data arrives
What
Fix SSE read timeout to 60 seconds instead of int max. Also fix user agent for SSE sockets.

Why
Manual testing of the iOS SDK has shown that the SSE event stream will not time out.
This means if a network connection dies and becomes stale we won’t detect this and switch over to polling, which will cause the cache to get out of sync.

Testing
Manual testing with proxy to confirm polling is restarted, SSE is reconnected