FFM-8160 - Add refreshCache() API method (#26)
* FFM-8160 - Add refreshCache() API method to allow apps coming to foreground to get latest flag states

What
Adds a new function called refreshCache() to CfClient

Why
This change adds a new API called refreshCache() which can be called by a mobile app coming to the foreground
to update any SSE flag events that were missed while the app was suspended. Note this new API should not be used to
poll the Feature Flags server, it will only retrive evaluations from the server if enough time has elapsed to
avoid putting to much pressure on the backend.