[FFM-9282] - Fix runtime failure when target is nil
What
Change target from implicitly-unwrapped optional to non-optional to avoid nil dereference errors.
Also set the target to a default value in the constructor and add some additional isValid()
checks to avoid cache lookups when the target is not yet known.

Why
SDK can throw an exception when initialization has not completed and boolVariation is called,
this is because target has not been set yet.

Testing
New unit test + manual